gpt 추가

+++ tripPlanner/src/main/java/com/example/tripPlanner/service/ReviewServiceImp.java
@@ -54,8 +54,6 @@ public class ReviewServiceImp implements ReviewService {
 			if (!placeService.exist(review.getPlaceId())) {
 				return false;
 			}
-			// 태그 달아주기
-			placeService.addTag(review.getPlaceId(), review.getTag());
 			// 여행지 DB에 있는 종합 평점에 반영
 			placeService.scorePlace(review.getPlaceId(), review.getScore());
 		}
@@ -64,23 +62,31 @@ public class ReviewServiceImp implements ReviewService {
 	}
 
 	@Override
-	public boolean deleteReview(Integer reviewId) {
+	public boolean deleteReview(Integer reviewId, String memberId) {
 
 		// 리뷰 객체 가져오기
 		Review review = reviewDao.getReview(reviewId);
 
 		if(review != null) {
-			// 삭제하려는 리뷰가 음식점 리뷰일 경우
-			if(review.getPlaceId().startsWith("R")) {
-				// 음식점 DB에 있는 종합 평점에 반영
-				restaurantService.cancelScore(review.getPlaceId(), review.getScore());
-				return reviewDao.deleteReview(reviewId);
-			}
-			// 삭제하려는 리뷰가 여행지 리뷰일 경우
-			else {
-				// 여행지 DB에 있는 종합 평점에 반영
-				placeService.cancelScore(review.getPlaceId(), review.getScore());
-				return reviewDao.deleteReview(reviewId);
+			
+			// 삭제하려는 리뷰의 작성자 아이디와 현재 접속한 아이디가 같다면
+			if(review.getMemberId().equals(memberId)) {
+				
+				// 삭제하려는 리뷰가 음식점 리뷰일 경우
+				if(review.getPlaceId().startsWith("R")) {
+					// 음식점 DB에 있는 종합 평점에 반영
+					restaurantService.cancelScore(review.getPlaceId(), review.getScore());
+					return reviewDao.deleteReview(reviewId);
+				}
+				// 삭제하려는 리뷰가 여행지 리뷰일 경우
+				else {
+					// 여행지 DB에 있는 종합 평점에 반영
+					placeService.cancelScore(review.getPlaceId(), review.getScore());
+					return reviewDao.deleteReview(reviewId);
+				}
+				
+			} else {
+				return false;
 			}
 		} else {
 			return false;
